+++ test/index.test.js
@@ -6,7 +6,7 @@ describe('App definition', () => {
     expect(typeof App.platformVersion).toBe('string')
     expect(App.authentication.type).toBe('oauth2')
     expect(App.hydrators.downloadSubmissionPdf).toBeDefined()
-    expect(Object.keys(App.triggers).sort()).toEqual(['form_list', 'public_link_submission', 'request_canceled', 'request_completed', 'request_expired'])
+    expect(Object.keys(App.triggers).sort()).toEqual(['form_list', 'public_link_submission', 'public_link_submission_abandoned', 'public_link_submission_updated', 'request_canceled', 'request_completed', 'request_expired'])
     expect(App.triggers.form_list.display.hidden).toBe(true)
     expect(Object.keys(App.creates).sort()).toEqual(['cancel_request', 'create_request', 'get_request', 'remind_request'])
     expect(Object.keys(App.searches)).toEqual(['find_request'])